add test for hexalines #55

+++ extract/extract.go
@@ -52,12 +52,12 @@ func Theme(r *http.Request) string {
 
 // Hexalines return the value of the hexalines parameter in the http.Request.
 // possible values: 6 or 8. Default value : 6
-func Hexalines(r *http.Request) int64 {
+func Hexalines(r *http.Request) int {
 	s := strings.ToLower(r.FormValue("hexalines"))
 	if len(s) > 0 {
 		if n, err := strconv.ParseInt(s, 0, 64); err == nil {
 			if n%6 == 0 || n%4 == 0 {
-				return n
+				return int(n)
 			}
 		}
 	}
